Move Screen.kt to Ion, and tab some spaces
